significant improvement to network editor
1. adding a new node on top of another node replaces that node. While we
want to keep this feature, we should have a warning when a node other
than a junction is being replaced.

2. deleting a node results in keeping existing links and replacing the
deleted node with a junction instead. Of course, this needs to be
revised so that if a junction is deleted, then it and associated links
are also deleted.

So, we can think of a junction as a base node, upon which other nodes
can be added or deleted.